refactor(routes): standardize project routes to plural form
