+++ pages/manage/projects/detail.tsx
