+++ pages/manage/projects/notes/index.tsx
@@ -1,20 +1,19 @@
 import toast from 'react-hot-toast';
 import { useMemo, useState } from 'react';
 import { useRouter } from 'next/router';
+import { ContentCard } from '@/features/projects';
 import getProjectLayout from '@/layout/ProjectLayout';
 import Button from '@/shared/components/Button';
-import ReviewCard from '@/components/Review/Card';
-import CreateModal from '@/components/Review/Modals/CreateModal';
 import SEOConfig from '@/shared/components/SEO';
-import UpdateModal from '@/components/Review/Modals/UpdateModal';
+import CreateModal from '@/components/Note/Modals/CreateModal';
+import UpdateModal from '@/components/Note/Modals/UpdateModal';
+import ConfirmModal from '@/shared/components/Confirm';
 import {
   useProject,
-  useProjectReview,
-  useProjectReviews,
-  useProjectReviewMutation,
+  useProjectNote,
+  useProjectNoteMutation,
+  useProjectNotes,
 } from '@/services/modules/projects';
-import ConfirmModal from '@/shared/components/Confirm';
-import marathonConfig from '@/constants/marathon';
 import { parseToString } from '@/services/core';
 
 enum ModalTypeEnum {
@@ -23,44 +22,46 @@ enum ModalTypeEnum {
   Delete,
 }
 
-const ReviewPage = () => {
+const NotesPage = () => {
   const { query } = useRouter();
   const projectId = parseToString(query.id);
   const [modalType, setModalType] = useState<ModalTypeEnum | null>(null);
-  const [reviewId, setReviewId] = useState<number | undefined>(undefined);
+  const [noteId, setNoteId] = useState<number | null>(null);
   const { data: project } = useProject(projectId);
 
-  const { data: reviews, mutate } = useProjectReviews(projectId);
+  const { data: notes, mutate } = useProjectNotes(projectId);
 
-  const { data: detail } = useProjectReview({
+  const { data: detail } = useProjectNote({
     projectId,
-    reviewId,
+    noteId,
   });
 
   const { createMutation, updateMutation, deleteMutation } =
-    useProjectReviewMutation({
+    useProjectNoteMutation({
       projectId,
-      reviewId,
+      noteId,
       onCreated: () => {
         toast.success('新增成功');
         setModalType(null);
+        mutate();
       },
       onUpdated: () => {
         toast.success('更新成功');
         setModalType(null);
-        setReviewId(undefined);
+        setNoteId(null);
         mutate();
       },
       onDeleted: () => {
         toast.success('刪除成功');
         setModalType(null);
-        setReviewId(undefined);
+        setNoteId(null);
+        mutate();
       },
     });
 
   const SEOData = useMemo(
     () => ({
-      title: `${project?.title} 覆盤｜島島阿學`,
+      title: `${project?.title} 便利貼｜島島阿學`,
       description:
         project?.description?.substring(0, 150) ||
         '「島島阿學」盼能透過建立多元的學習資源網絡，讓自主學習者能找到合適的成長方法，進一步成為自己想成為的人，從中培養共好精神。目前正積極打造「可共編的學習資源平台」。',
@@ -68,7 +69,7 @@ const ReviewPage = () => {
       author: '島島阿學',
       copyright: '島島阿學',
       imgLink: 'https://www.daoedu.tw/preview.webp',
-      link: `${process.env.HOSTNAME}/manage/project/review?id=${projectId}`,
+      link: `${process.env.HOSTNAME}/manage/projects/notes?id=${projectId}`,
     }),
     [project?.title, project?.description, projectId]
   );
@@ -80,76 +81,75 @@ const ReviewPage = () => {
   return (
     <>
       <SEOConfig data={SEOData} />
-      <div className="mb-6 flex items-end sm:items-center justify-between body-md">
-        <div className="flex flex-col items-start sm:flex-row sm:items-center gap-1">
-          <div className="text-basic-500">
-            覆盤（{marathonConfig.getWeekNumber().toString().padStart(2, '0')}{' '}
-            週/22週）
-          </div>
-        </div>
+      <div className="mb-6 flex items-center justify-between body-md">
+        <div className="text-basic-500">便利貼 ({notes?.length || 0})</div>
         <Button
           variant="solid"
           color="primary"
           onClick={() => setModalType(ModalTypeEnum.Create)}
         >
-          新增覆盤
+          新增便利貼
         </Button>
       </div>
-
-      <ul className="flex flex-col gap-6">
-        {Array.isArray(reviews) &&
-          reviews.map((review) => (
-            <li key={review.id}>
-              <ReviewCard
-                data={review}
-                detailLink={`/manage/project/reviews/detail?id=${projectId}&reviewId=${review.id}`}
-                onEditClick={() => {
-                  setReviewId(review.id);
-                  setModalType(ModalTypeEnum.Update);
-                }}
-                onDeleteClick={() => {
-                  setReviewId(review.id);
-                  setModalType(ModalTypeEnum.Delete);
-                }}
-              />
-            </li>
-          ))}
+      <ul className="px-4 bg-basic-white flex flex-col rounded-2xl">
+        {notes?.map((note) => (
+          <li
+            key={note.id}
+            className="py-6 border-b last:border-b-0 border-solid border-basic-200"
+          >
+            <ContentCard
+              type="note"
+              data={note}
+              className="p-3 transition-shadow hover:shadow-basic-200/40 hover:shadow-lg"
+              detailLink={`/manage/projects/notes/detail?id=${projectId}&noteId=${note.id}`}
+              onEditClick={() => {
+                setModalType(ModalTypeEnum.Update);
+                setNoteId(note.id);
+              }}
+              onDeleteClick={() => {
+                setModalType(ModalTypeEnum.Delete);
+                setNoteId(note.id);
+              }}
+            />
+          </li>
+        ))}
       </ul>
 
       {project && (
         <CreateModal
-          projectId={projectId}
-          projectTitle={project.title}
           isOpen={modalType === ModalTypeEnum.Create}
           onClose={() => setModalType(null)}
-          onSubmit={createMutation.trigger}
-          isLoading={createMutation.isMutating}
-        />
-      )}
-
-      {detail && reviewId && project && (
-        <UpdateModal
           projectId={projectId}
           projectTitle={project.title}
-          reviewId={reviewId}
-          defaultValues={detail}
-          week={detail.week}
-          createdAt={detail.createdAt}
-          isOpen={modalType === ModalTypeEnum.Update}
-          onClose={() => setModalType(null)}
-          onSubmit={updateMutation.trigger}
-          isLoading={updateMutation.isMutating}
+          isLoading={createMutation.isMutating}
+          onSubmit={createMutation.trigger}
         />
       )}
 
-      {reviewId && (
+      {detail && noteId && project && (
+        <UpdateModal
+          key={noteId}
+          id={noteId}
+          isOpen={modalType === ModalTypeEnum.Update}
+          onClose={() => setModalType(null)}
+          projectId={projectId}
+          projectTitle={project.title}
+          week={detail.week}
+          createdAt={detail.date}
+          isLoading={updateMutation.isMutating}
+          defaultValues={detail}
+          onSubmit={updateMutation.trigger}
+        />
+      )}
+
+      {noteId && (
         <ConfirmModal
-          title="確認刪除覆盤"
+          title="確認刪除便利貼"
           confirmText="確認刪除"
           confirmColor="alert"
           isOpen={modalType === ModalTypeEnum.Delete}
           onClose={() => setModalType(null)}
-          onConfirm={() => deleteMutation.trigger({ projectId, reviewId })}
+          onConfirm={() => deleteMutation.trigger({ projectId, noteId })}
           isLoading={deleteMutation.isMutating}
         />
       )}
@@ -157,6 +157,6 @@ const ReviewPage = () => {
   );
 };
 
-ReviewPage.getLayout = getProjectLayout;
+NotesPage.getLayout = getProjectLayout;
 
-export default ReviewPage;
+export default NotesPage;
